requirements.txt
anyio==4.8.0
black==25.1.0
blessed==1.20.0
certifi==2025.1.31
click==8.1.8
codefind==0.1.7
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jurigged==0.6.0
loguru==0.7.3
mypy-extensions==1.0.0
ovld==0.4.6
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
prettytable==3.15.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-telegram-bot==21.10
six==1.17.0
sniffio==1.3.1
watchdog==6.0.0
wcwidth==0.2.13
